Bangalore Housing Prediction Model using Scikit-Learn
From the Kaggle dataset, I have downloaded Bangalore Hosing Dataset to practice creating machine learning model that can predict the price of the houses based on the features. I performed preprocessing on the dataset such as standardizing, cleaning , and removing outliers and unreliable data so that it can be trained using scikit-learn modules.